Break out Parallax out of ParallaxComponent
