+++ doc/components.md
@@ -197,7 +197,7 @@ The simplest `ParallaxComponent` is created like this:
 ```dart
   @override
   Future<void> onLoad() async {
-    final parallax = await ParallaxComponent.load(
+    final parallaxComponent = await ParallaxComponent.load(
       [bg.png, trees.png],
     );
     add(parallax);
@@ -212,7 +212,7 @@ in the `load` helper function.
 For example if you want to move your background images along the X-axis with a faster speed the "closer" the image is:
 
 ```dart
-  final parallax = await ParallaxComponent.load(
+  final parallaxComponent = await ParallaxComponent.load(
     _paths,
     baseVelocity: Vector2(20, 0),
     velocityMultiplierDelta: Vector2(1.8, 1.0),
@@ -221,8 +221,9 @@ For example if you want to move your background images along the X-axis with a f
 You can set the baseSpeed and layerDelta at any time, for example if your character jumps or your game speeds up.
 
 ```dart
-  parallaxComponent.baseSpeed = Vector2(100, 0);
-  parallaxComponent.velocityMultiplierDelta = Vector2(2.0, 1.0);
+  final parallax = parallaxComponen.parallax;
+  parallax.baseSpeed = Vector2(100, 0);
+  parallax.velocityMultiplierDelta = Vector2(2.0, 1.0);
 ```
 
 By default the images are aligned to the bottom left, repeated along the X-axis and scaled proportionally so that
@@ -238,7 +239,12 @@ Advanced example:
     ParallaxImage.load('dust.jpg', repeat: ImageRepeat.repeatX, alignment: Alignment.topRight, fill: LayerFill.height),
   ];
   final layers = images.map((image) => ParallaxLayer(await image, velocityMulitplier: images.indexOf(image) * 2.0));
-  final parallaxComponent = ParallaxComponent(await Future.wait(layers), baseVelocity: Vector2(50, 0));
+  final parallaxComponent = ParallaxComponent(
+    Parallax(
+      await Future.wait(layers),
+      baseVelocity: Vector2(50, 0),
+    ),
+  );
 ```
 
 * The stars image in this example will be repeatedly drawn in both axis, align in the center and be scaled to fill the screen width.
